refactor create_remote and df into module in CLI
