+++ src/cli/src/cmd_setup.rs
@@ -7,6 +7,7 @@ use liboxen::command::migrate::{
 use liboxen::constants::{DEFAULT_BRANCH_NAME, DEFAULT_REMOTE_NAME};
 
 use crate::cmd::add::add_args;
+use crate::cmd::df::DFCmd;
 use crate::cmd::remote::commit::RemoteCommitCmd;
 
 pub const ADD: &str = "add";
@@ -41,44 +42,6 @@ pub const SCHEMAS: &str = "schemas";
 pub const STATUS: &str = "status";
 pub const UPLOAD: &str = "upload";
 
-pub fn create_remote() -> Command {
-    Command::new(CREATE_REMOTE)
-        .about("Creates a remote repository with the name on the host. Default behavior is to create a remote on the hub.oxen.ai remote.")
-        .arg(
-            Arg::new("name")
-                .long("name")
-                .short('n')
-                .help("The namespace/name of the remote repository you want to create. For example: 'ox/my_repo'")
-                .required(true)
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("host")
-                .long("host")
-                .help("The host you want to create the remote repository on. For example: 'hub.oxen.ai'")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("scheme")
-                .long("scheme")
-                .help("The scheme for the url of the remote repository. For example: 'https' or 'http'")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("add_readme")
-                .long("add_readme")
-                .help("If present, it will create a README file and initial commit in the remote repo.")
-                .action(clap::ArgAction::SetTrue),
-        )
-        .arg(
-            Arg::new("is_public")
-                .long("is_public")
-                .short('p')
-                .help("If present, it will create a public remote repository.")
-                .action(clap::ArgAction::SetTrue),
-        )
-}
-
 pub fn remote() -> Command {
     Command::new(REMOTE)
         .about("Interact with a remote repository without cloning everything locally.")
@@ -93,7 +56,7 @@ pub fn remote() -> Command {
                 .action(clap::ArgAction::Set))
         )
         .subcommand(RemoteCommitCmd.args())
-        .subcommand(df())
+        .subcommand(DFCmd.args())
         .subcommand(diff())
         .subcommand(download())
         .subcommand(log())
@@ -238,201 +201,6 @@ pub fn ls() -> Command {
         )
 }
 
-pub fn df() -> Command {
-    Command::new(DF)
-        .about("View and transform data frames. Supported types: csv, tsv, ndjson, jsonl, parquet.")
-        .arg(arg!(<DF_SPEC> ... "The DataFrame you want to process. If in the schema subcommand the schema ref."))
-        .arg_required_else_help(true)
-        .arg(
-            Arg::new("output")
-                .long("output")
-                .short('o')
-                .help("Output file to store the transformed data")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("delimiter")
-                .long("delimiter")
-                .short('d')
-                .help("The delimiter to use when reading the file. Default is ','")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("columns")
-                .long("columns")
-                .short('c')
-                .help("A comma separated set of columns names to look at. Ex file,x,y")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("filter")
-                .long("filter")
-                .short('f')
-                .help("An filter the row data based on an expression. Supported Ops (=, !=, >, <, <= , >=) Supported dtypes (str,int,float)")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("aggregate")
-                .long("aggregate")
-                .short('a')
-                .help("Aggregate up values based on field.")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("col-at")
-                .long("col-at")
-                .help("Select a specific row item from column to view it fully. Format: 'col_name:index' ie: 'my_col_name:3'")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("row")
-                .long("row")
-                .help("Select a specific row to view it fully. Format: '3'")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("vstack")
-                .long("vstack")
-                .help("Combine row data from different files. The number of columns must match.")
-                .action(clap::ArgAction::Append),
-        )
-        .arg(
-            Arg::new("slice")
-                .long("slice")
-                .help("A continuous slice of the data you want to look at. Format: 'start..end' Ex) '10..25' will take 15 elements, starting at 10 and ending at 25.")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("head")
-                .long("head")
-                .help("Grab the first N entries of the data frame.")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("tail")
-                .long("tail")
-                .help("Grab the last N entries of the data frame.")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("page")
-                .long("page")
-                .help("Page number when paginating through the data frame. Default page = 1")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("page-size")
-                .long("page-size")
-                .help("Paginated through the data frame. Default page-size = 10")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("take")
-                .long("take")
-                .short('t')
-                .help("A comma separated set of row indices to look at. Ex 1,22,313")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("index")
-                .long("index")
-                .help("Indexes the dataframe for editing via Oxen's remote staging environment.")
-                .action(clap::ArgAction::SetTrue),
-        )
-        .arg(
-            Arg::new("add-col")
-                .long("add-col")
-                .help("Add a column with a default value to the data table. If used with --add-row, row is added first, then column. Format 'name:val:dtype'")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("add-row")
-                .long("add-row")
-                .help("Add a row and cast to the values data types to match the current schema. If used with --add-col, row is added first, then column. Format 'comma,separated,vals'")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("delete-row")
-                .long("delete-row")
-                .help("Delete a row from a data frame. Currently only works with remote data frames with the value from _id column.")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("get-row")
-                .long("get-row")
-                .help("Retrieve a specific row from the data frame by row id.")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("content-type")
-                .long("content-type")
-                .help("The data that you want to append to the end of the file. Valid content types are 'json', 'csv', 'text'.")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("sort")
-                .long("sort")
-                .short('s')
-                .help("Sort the output by a column name. Is run at the end of all the other transforms.")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("sql")
-                .long("sql")
-                .help("Run a sql query on the data frame.")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("committed")
-                .long("committed")
-                .help("Show the committed version of the dataframe on the current branch, not any remotely staged data.")
-                .action(clap::ArgAction::SetTrue),
-        )
-        .arg(
-            Arg::new("text2sql")
-                .long("text2sql")
-                .help("Run a text query that translates to sql on the data frame.")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("host")
-                .long("host")
-                .help("What remote host to run the query against. Ie: hub.oxen.ai")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("unique")
-                .long("unique")
-                .short('u')
-                .help("Unique the output by a set of column names. Takes a comma separated set of column names ie: \"text,label\".")
-                .action(clap::ArgAction::Set),
-        )
-        .arg(
-            Arg::new("randomize")
-                .long("randomize")
-                .help("Randomize the order of the table")
-                .action(clap::ArgAction::SetTrue),
-        )
-        .arg(
-            Arg::new("reverse")
-                .long("reverse")
-                .help("Reverse the order of the table")
-                .action(clap::ArgAction::SetTrue),
-        )
-        .arg(
-            Arg::new("schema")
-                .long("schema")
-                .help("Print the full list of columns and data types within the schema in a dataframe.")
-                .action(clap::ArgAction::SetTrue),
-        )
-        .arg(
-            Arg::new("schema-flat")
-                .long("schema-flat")
-                .help("Print the full list of columns and data types within the schema.")
-                .action(clap::ArgAction::SetTrue),
-        )
-}
-
 pub fn schemas() -> Command {
     Command::new(SCHEMAS)
         .about("Manage schemas that are created from committing tabular data")
@@ -500,7 +268,7 @@ pub fn schemas() -> Command {
                 .arg(Arg::new("METADATA").help("Any additional metadata you want to add to the schema."))
 
         )
-        .subcommand(df())
+        .subcommand(DFCmd.args())
 }
 
 pub fn download() -> Command {
