+++ src/lib/Cargo.toml
@@ -33,7 +33,7 @@ bytevec = "0.2.0"
 chrono = "0.4.22"
 clap = { version = "4.2.7", features = ["cargo"] }
 colored = "2.0.0"
-comfy-table = "7.0.0"
+comfy-table = "7.1.1"
 deadqueue = "0.2.4"
 derive_more = "0.99.17"
 difference = "2.0.0"
